@@ -2,19 +2,22 @@
   "name": "viz",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": ">=18"
+  },
   "dependencies": {
-    "@rehooks/window-size": "^1.0.2",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "p5": "^1.11.3",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-draggable": "^4.3.1",
-    "react-p5": "^1.3.2",
-    "react-scripts": "3.4.1",
-    "react-tabs": "^3.1.0",
-    "use-persisted-state": "^0.3.0",
+    "react-p5": "^1.4.1",
+    "react-scripts": "^5.0.1",
+    "react-tabs": "^6.1.0",
     "uuid": "^8.0.0"
   },
   "scripts": {
