@@ -2,8 +2,11 @@ import { render } from '@testing-library/react';
 import React from 'react';
 import App from './App';
 
-test('renders learn react link', () => {
-  const { getByText } = render(<App />);
-  const linkElement = getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+jest.mock('./P5Viz/P5Viz', () => () => <div data-testid="p5viz" />);
+jest.mock('./Controls/Controls', () => () => <div data-testid="controls" />);
+
+test('renders the app shell', () => {
+  const { getByTestId } = render(<App />);
+  expect(getByTestId('p5viz')).toBeInTheDocument();
+  expect(getByTestId('controls')).toBeInTheDocument();
 });
